requirements.txt
fastapi 
uvicorn 
pandas 
anthropic
gspread
oauth2client
python-dotenv